refactor: changed data.values to Map
this allows for the setting of "column" names, which could be displayed
on the axes or as information on hover etc.
The special name "groups" is used for setting the groups for the values.
Still works with xValues, yValues and groups.
Moved styling, grouping and setting inital data to the beginning of
plot, so it does not recalculate evrytime validateComponentData is
called.
Added tests